refactor: types 파일로 타입 정의 이동
